refactor: replace lang3 StringUtils use as well
