+++ client/oslc-client/src/main/java/org/eclipse/lyo/client/OslcOAuthClient.java
@@ -21,8 +21,8 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.http.client.HttpClient;
+import org.eclipse.lyo.core.util.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -58,7 +58,7 @@ public class OslcOAuthClient implements IOslcClient {
 
         oauthRealmName = "Jazz";
         // Change if a different name was detected
-        if (!StringUtils.isEmpty(realm)) {
+        if (!StringUtils.isNullOrEmpty(realm)) {
             oauthRealmName = realm;
         }
