+++ core/lyo-core-settings/src/main/java/org/eclipse/lyo/core/util/StringUtils.java
@@ -1,8 +1,27 @@
-package org.eclipse.lyo.oslc4j.core;
+/*
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0, or the Eclipse Distribution License 1.0
+ * which is available at http://www.eclipse.org/org/documents/edl-v10.php.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
+ */
+package org.eclipse.lyo.core.util;
 
 import java.util.regex.Pattern;
 
+/**
+ * @since 7.0.0
+ */
 public class StringUtils {
+    /**
+     * Pattern to match control characters in the Unicode Cc category that are not CR, LF, or TAB
+     */
     private static final Pattern CONTROL_CHAR_PATTERN = Pattern.compile("^\\p{Cc}&&[^\\r\\n\\t]+$");
 
     /**
@@ -26,7 +45,6 @@ public class StringUtils {
     public static boolean isNullOrWhitespace(String str) {
         return str == null || str.isBlank();
     }
-
 
     public static boolean isNullOrEmpty(String str) {
         return str == null || str.isEmpty();
