+++ server/oslc-ui-model/pom.xml
@@ -25,10 +25,6 @@
           <groupId>com.fasterxml.jackson.core</groupId>
           <artifactId>jackson-core</artifactId>
         </dependency>
-        <dependency>
-          <groupId>org.apache.commons</groupId>
-          <artifactId>commons-lang3</artifactId>
-        </dependency>
     </dependencies>
     <build>
       <plugins>
